move files a bit
